classe Menu ajustada, agora classe Main apenas chama o método executaMenu()
